==> Project files/1. Read Country IDs.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import re

#Country Codes
country_ids_G4 = pd.read_excel('Country Code/T19_G4_Curriculum Questionnaire Data.xlsx', usecols=[0,1],
                                 names=['Country', 'country_ID'], skiprows=[1,2,61], skipfooter=94)

country_ids_G8 = pd.read_excel('Country Code/T19_G8_Curriculum Questionnaire Data.xlsx', usecols=[0,1], 
                             names=['Country', 'country_ID'], skiprows=[1,2,42], skipfooter=113)

# add grade columns to both dataframes
country_ids_G4['grade'] = 'yes'
country_ids_G8['grade'] = 'yes'

# join all country code dataframes
country_ids = pd.merge(country_ids_G4, country_ids_G8, how='outer', 
                         on=['Country', 'country_ID'])

# eliminate parenthesis and replace country names
country_ids['Country'] = country_ids['Country'].replace('[\(].*[\)]', '', regex=True)

to_replace = ['Moscow City, Russian Fed.', 'Russian Federation', 'Chinese Taipei','Hong Kong SAR', 
              'Iran, Islamic Rep. of', 'Korea, Rep. of', 'Ontario, Canada', 'Quebec, Canada', 
              'Abu Dhabi, UAE', 'Dubai, UAE', 'Czech Republic', 'Slovak Republic', 'Belgium ']
replace_with = ['Russia, Moscow', 'Russia', 'Taiwan', 'Hong Kong', 
                'Iran', 'South Korea', 'Canada, Ontario', 'Canada, Quebec', 
                'United Arab Emirates, Abu Dhabi', 'United Arab Emirates, Dubai', 'Czechia', 'Slovakia', 'Belgium']                                                 

country_ids['Country'] = country_ids['Country'].replace(to_replace, replace_with)

# rename grade columns and replace NaNs with 'no'
country_ids = country_ids.rename(columns={'grade_x':'grade_4', 'grade_y':'grade_8'}) \
                         .fillna('no').sort_values('Country')

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

country_ids

country_ids.to_csv('country_ids.csv', index=True)
